# docs/openapi.yaml
openapi: 3.0.3
info:
  title: CineAI API
  description: Comprehensive movie recommendation API with AI-powered features
  version: 1.0.0
  contact:
    name: CineAI Team
    email: contact@cineai.app
  license:
    name: MIT
servers:
  - url: https://cineai.vercel.app/api
    description: Production server
  - url: http://localhost:3000/api
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Movie:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        overview:
          type: string
        posterPath:
          type: string
        backdropPath:
          type: string
        releaseDate:
          type: string
          format: date
        runtime:
          type: integer
        rating:
          type: number
          format: float
        voteCount:
          type: integer
        genres:
          type: array
          items:
            type: string
        cast:
          type: array
          items:
            type: object
        confidence:
          type: number
          format: float
          minimum: 0
          maximum: 1
...
paths:
...
  /movies/search:
    post:
      summary: Search movies with natural language
      tags:
        - Movies
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                query:
                  type: string
                  example: "sci-fi movies like Blade Runner"
                limit:
                  type: integer
                  minimum: 1
                  maximum: 50
                  default: 10
              required:
                - query
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      movies:
                        type: array
                        items:
                          $ref: '#/components/schemas/Movie'
...
  /recommendations/hyper-personalized:
    get:
      summary: Get hyper-personalized recommendations
      tags:
        - AI & Recommendations
      security:
        - bearerAuth: []
      parameters:
        - name: count
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
        - name: context
          in: query
          schema:
            type: string
        - name: excludeWatched
          in: query
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: Personalized recommendations
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      recommendations:
                        type: array
                        items:
                          $ref: '#/components/schemas/Movie'
...
security:
  - bearerAuth: []